= SipStack =  * doxygen 을 위한 주석을 추가함

+++ trunk/SipStack/TcpConnectMap.cpp
@@ -158,7 +158,7 @@ int CTcpConnectMap::GetSize()
 	int iSize;
 
 	m_clsMutex.acquire();
-	iSize = m_clsMap.size();
+	iSize = (int)m_clsMap.size();
 	m_clsMutex.release();
 
 	return iSize;
